+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Changed to Federated plugins
 - Support Web(bootForWeb)
 - Changed `updateUser` API's parameters
+- Added `openSupportBot`, `setAppearance` API
 
 ## 2.6.2
 - Fixed iOS ChannelIOSdk version to 11.6.0